Two ("three") API tests added in Chirp.ChirpWeb.Tests
Two are running and work (get public timeline, get private timeline), but one is commented out as data seems not to be transfered to the instance we use for testing. I can look at this more later, but think we might have other tests later that will also test this, so maybe unnescesary work.

Co-authored-by: Sophie Gabriela <slak@itu.dk>