Debugging changes to LINQ statement in AuthorRepo
Fixing of buttons in public.cshtml to only show up when user logged in.

Changed in Author.cs to override IdentityUser UserName and make it required - idk if we should do this? Did not update the migration to reflect it.

Current error: Upon login problems in register class due to SQLITE error 1: a.AuthorId column does not exist.

Co-authored-by: Sophie Gabriela <slak@itu.dk>